Add AuthenticationResultBuilder (#7301)
